requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
transformers==4.24.0
Pillow==9.0.0
torch==1.13.1
torchvision==0.14.1
sentencepiece==0.1.97
protobuf==3.20.0
python-multipart==0.0.6
